// src/app/api/models/asiento-circunscripcion-insert-request.ts
/* tslint:disable */
/* eslint-disable */
export interface AsientoCircunscripcionInsertRequest {
  circunscripcionDestinos?: null | string;
  circunscripcionOrigenes?: null | string;
  codTipoAsiento?: number;
  descripcion?: null | string;
  detallesModificacion?: null | string;
  fechaAnotacion?: null | string;
  idDispositivo?: null | number;
  idInformeRenac?: null | number;
  idTipoAsiento?: null | number;
  informacionComplementaria?: null | string;
  nombreCapital?: null | string;
  nombreCircunscripcion?: null | string;
  nombreDepartamento?: null | string;
  nombreProvincia?: null | string;
  numeroAsiento?: null | string;
  tipoCircunscripcion?: null | string;
}