// src/app/api/models/asiento-circunscripcion-update-request.ts
/* tslint:disable */
/* eslint-disable */
export interface AsientoCircunscripcionUpdateRequest {
  circunscripcionDestinos?: null | string;
  circunscripcionOrigenes?: null | string;
  codTipoAsiento?: number;
  descripcion?: null | string;
  detallesModificacion?: null | string;
  fechaAnotacion?: null | string;
  idAsientoCircunscripcion?: null | number;
  idDispositivo?: null | number;
  idInformeRenac?: null | number;
  idTipoAsiento?: null | number;
  informacionComplementaria?: null | string;
  nombreCapital?: null | string;
  nombreCircunscripcion?: null | string;
  nombreDepartamento?: null | string;
  nombreProvincia?: null | string;
  numeroAsiento?: null | string;
  tipoCircunscripcion?: null | string;
}